Moving form dialog to its own file at @jacalz request
 - Renamed test file to make it clear what's being tested.
 - Started on FormDialog struct
